gtk: disconnect record stream when record_stop()
This behaviour is different than playback. The main difference is that
a playback stream will stay active, while the recording stream will be
removed: that way the GNOME volume control can notice the user there
is no active recording, by hiding the microphone icon.

Playback is more frequent (sounds events...) and it would be
unfriendly if the stream was removed/added every now and then from the
volume control list.