Adapt build system to allow building with GTK3
The new configure flag '--with-gtk' can be used to choose
which GTK version to build against, defaulting to GTK2.
To enable GTK3 use

   ./configure --with-gtk=3.0

The libspice-client-glib-2.0.la library is unchanged, building
against glib-2.0 at all times.

The GTK3 build will produce a libspice-client-gtk-3.0.la
The include files will also live in $prefix/spice-client-gtk-3.0
and the pkgconfig is called spice-client-gtk-3.0 too.

This allows for full parallel install of GTK2 and GTK3 builds