spicy: disable Ctrl+W/Close keyboard shortcut
The keyboard shortcut for the Close gtk action isn't disabled,
which leads to spicy closing when trying to use Ctrl+W in the guest.
All other problematic shortcuts are disabled, but in this case the
disabling was commented out. Reenable it to prevent people from
accidentally closing spicy when pressing ctrl+w