factor base message handling in SpiceChannel
Currently, every channel has to define all the server messages it
handles, including the "generic" ones. Since it's error prone (easy
to forget the handling of default messages in new sources), it's
better to move this handling to the base channel class, and to call
the parent method when the message is unknown in the ::handle_msg
method.

On top of this, another factoring that can be done is to make the
message handling function generic instead of reimplementing it in
every class. Each class would only have to register its own
(class-specific) set of handlers.

Conflicts:

	gtk/channel-playback.c
	gtk/channel-record.c