add basic callbacks for smartcard events
Now that the smartcard manager sends event for smartcard readers,
the smartcard channel has to do something with them. For now, we just
add basic callbacks which don't try to forward these events to the
spice server. However, we start adding the glue needed to handle
the fact that plugging of smartcard readers and creation of these
readers in the spice server will be asynchronous.
After receiving a "reader addition" command, the spice server sends
back an ID to identify this reader which must be used in all other
requests (card insertion, removal and reader removal) to identify
this reader. However, if the server is slow to send back this ID,
there's a window when we can get additional events for the reader
that is being added that we can't send right away to the server
because we don't know the ID to use. That's why we add the various
pending_* hashes in this commit, to be able to keep track of the
requests that will have to be sent once the spice server has
assigned an ID to the reader.