serialize sending of smartcard channel messages
Messages sent to the spice server by the smartcard channel have to
be serialized: before sending a message, the channel has to wait
for the server answer to the previous message. Add a GQueue to be
able to queue several messages to the server while we wait for the
answer to an earlier message.
I think the reason why that serialization is needed is to be able
to deterministically assign IDs to readers on reader addition. We
have no way to match a message from the spice server with the
message to the server that triggered this reply, which means that
if we could send several reader additions to the server without
waiting for answers, when the server answers, we wouldn't know
which reader it's trying to assign an ID to.