don't always serialize messages sent to the server
The communication between spice clients and servers on the smartcard
channel can be initiated either by the client or by the server.
It's initiated by the client for smartcard reader events (reader
hot(un)plug, card insertion/removal), or it can be initiated by the
server when it wants to query the certificates available on the
smartcard.
When communication is initiated by the client, we want to serialize
the messages we sent, ie we don't want to send a message if we
haven't received yet the answer to the previous message.
However, when it's the server which initiates the communication,
we don't want to use this serializing mechanism.

This commit adds a "serialize_msg" boolean to be able to disable
message sending serialization as needed.