add --smartcard option to spicy
People starting spicy will not always want their smartcard data
to be forwarded to the guest they're connecting to. Currently,
smartcard events are unconditionnally forwarded to the server if
spicy was compiled with smartcard support. This commit adds a
--smartcard option that must be used in order to enable smartcard
support in the client. By default, smartcard data won't be forwarded
to the server unless this option is specified.