mjpeg: restrict use of i and j
Using i and j as variable names that are used from one loop to the
other isn't really readable, and makes the code more fragile than
it could be. This commits adds a "lines_read" variable which is more
expressive than "j", restricts "j" lifetime to the loop where it's
used, and it removes the "i" variable and uses counters provided
by libjpeg to iterate all the image lines.
It also has the side-effect that if jpeg_read_scanlines returns a short
read (less lines than expected are read), "dest" won't go out of sync but
will be set to the right place at the end of the loop.