spice-channel: Reset SPICE_COMMON_CAP_PROTOCOL_AUTH_SELECTION on disconnect
Our disconnect handler clears the common_caps array so that a new
connection starts with a clean slate. But in our constructor we set
the SPICE_COMMON_CAP_PROTOCOL_AUTH_SELECTION in common_caps as starting
cap. Do the same on disconnect, so the behavior of a re-using a channel
after disconnect is the same as using a fresh channel.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>