Add a private copy of gusb
While working on usb redirection support for spice-gtk I needed some
code to integrate libusb into glib's mainloop amongst other things. I ended
up borrowing code from colord for this. Richard (the colord author) and I
quickly agreed that doing generic glib bindings for libusb is a good idea,
akin to the gudev bindings for libudev we've called our WIP on this gusb:
https://gitorious.org/gusb

Since this very much is a WIP, the API is nowere near stable, so for now
we bundle a copy of this code with spice-gtk. When gusb has an official
release out the door with a stable API we should switch to that.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>