spice-widget: release_keys on focus out
This fixes alt getting stuck in the guest when the user alt-tabs away from the
spice-widget (thus making it see the alt press but not the release) and then
closing it without giving it the focus back (by using the windows close button
for example).

Note that doing this on focus out (when we know we may be missing key releases
afterwards) makes a lot more sense then doing this on focus in, and with
the release_keys on focus out I don't really see a reason any more to do it
on focus in. But I'm leaving it in on focus in just in case, it certainly
cannot hurt there.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>