Remove spice-client-gtk.defs from git (as it is auto-generated)
spice-client-gtk.defs gets autogenerated (but only when building for
gtk2). Having it in git means having to keep it in sync with the public
API manually, which requires doing a gtk2 build before committing each
time one makes changes to the public API.

As with any autogenerated files, this really does not belong in git,
removing it from git removes the need for the manual syncing.

Also add it to make clean, and change gtk/Makefile.am so that
the contents on EXTRA_DIST don't depend on how ./configure was run
making "make dist"-s result depend on ./configure flags is not a good idea.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>