gtk/Makefile.am: Don't include auto-generated files into make dist tarbal
We have various files which are auto-generated which currently get included
into the tarbal, yet they are part of CLEANFILES, so the first make clean
removes them, which is rather weird. Use nodist_foo_SOURCES to make them
not end up in the tarbal.

generated_[de]marshallers[1].c, are special as autogenerating those on
end user builds would mean requiring the end user to have pyparser installed,
so we add them to EXTRA_DIST and remove them from CLEANFILES. It may seem
weird to have them in nodist_... and then add them to EXTRA_DIST,
but this patch also adds an explict depenency on the autogenerated files
to the non autogerated ones:
$(libspice_client_glib_2_0_la_SOURCES): spice-glib-enums.h

Leaving generated_[de]marshallers[1].c in libspice_client_glib_2_0_la_SOURCES,
would cause it to get a dep on autogenerated files, and thus get regenerated
itself no matter what, defeating the purpose of having them in dist in the
first place.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>