SpiceDisplay: Pass on notify signal from GtkSession for GtkSession properties
Since some SpiceDisplay properties are now just a pass through to the
equivalent SpiceGtkSession property, we also need to pass on the notify
signal for these, so interested users can get notified of them changing.

Use this in spicy to ensure that changes to SpiceGtkSession (and thus per
session rather then per window) properties get reflected in all windows for
a multimonitor guest.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>