SpiceDisplay: Don't set auto-clipboard on construction
The auto-clipboard setting now lives in the SpiceGtkSession, if that
was changed to FALSE (from the default of TRUE) and later a SpiceDisplay
was created for this session that would change it back to TRUE again because
it would set it to its default from its constructor. This patch fixes this.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>