Correctly parse RFC-conform URIs separating the host and the arguments by a '/'
Little patch to make spice-gtk accept RFC-conform URLs of the form
"spice://host/?port=5901".

I didn't add '/' to the list of the other characters to ignore (?;&)
by intention since an URL like spice://host/?port=5901/somotherstuff
should not be valid, resp. the password may contain '/' because the
string is already unescaped at that point.

Unfortunately glib does not seem to have functions to 'explode' an URI
which would be really helpful.

Cheers,
Tiziano