channel-usbredir: Call the acl helper without first trying to open the device
Normally opening the USB device without first calling the helper will fail,
except when the process using spice-gtk is running as root.

So the current code which first tries to open the USB device before calling
the helper is optimizing for an exception rather then for the default code
path. More over it also causes libusb to print the following errors to stderr:
libusb:error [op_open] libusb couldn't open USB device /dev/bus/usb/002/017: Permission denied.
libusb:error [op_open] libusb requires write access to USB device nodes.

So this patch changes things to first call the helper and only then try
to open the device node.

This patch also modifies the helper to not call policykit when called by
a root process, since the set_facl which it will do, if policykit says it
is ok, is a no-op for root anyways. Instead it directly returns a success
status without doing anything when called by a root process.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>