usbredir: Use new libusbredirhost write flush callback
The new (in usbredir-0.3.2) usbredirhost_open_full()
function allows us to be notified whenever usb packets completing
result in data to be send to the host. This removes the need for using
g_usb_source_set_callback and seeing if there is data to write on
any of the usb channels each time some usb event happens.

This also bumps the minimum needed libusbredirhost version to 0.3.2

Signed-off-by: Hans de Goede <hdegoede@redhat.com>