Support semi-seamless migration
Yonit Halperin described the flow, that I followed:

(1) when client_migrate_info is called SPICE_MSG_MAIN_MIGRATE_BEGIN is
sent to the client.

Then, the client should link to the target server (SpiceLinkMess),
i.e., connect all the channels, but it shouldn't poll from the target,
only from the source. You can refer to RedClient::Migrate class.  The
connection id in the link message should be the id of the connection
to the source server.

(2) The client sends to the source server
SPICE_MSGC_MAIN_MIGRATE_(CONNECTED|CONNECT_ERROR)

(3) When migration completes  SPICE_MSG_MAIN_MIGRATE_(END|CANCEL) is
sent to the client.

(3.1) In case of SPICE_MSG_MAIN_MIGRATE_CANCEL, the client closes the
connections to the target.

(3.2) In case of SPICE_MSG_MAIN_MIGRATE_END, the client should reset
all the data that is related to the connection to the source and
complete the transition to the target server (without sending
ATTACH_CHANNELS, and without guest display initialization via agent).

Then, the client sends SPICE_MSG_MAIN_MIGRATE_END to the target.