usbredir: Add support for filtering out devices from auto-redirection
Note this patch adds a default filter filtering out HID devices, since
redirecting ie a mouse plugged into a laptop is usually not what the user
wants. Also sometimes usb keyboards/mice may experience a glitch causing
them to temporarily drop of the bus. If this happens when the spice-client
has focussed it used to redirect them to the vm when they re-appeared,
not good!

Signed-off-by: Hans de Goede <hdegoede@redhat.com>