Handle spice_audio_new failures
spice_audio_new can return a NULL pointer when there's a failure
during the initialization of the audio system. When this happens,
we shouldn't keep initializing the spice audio channel as if nothing
happened, but just stop the connection.
This can be tested by forcing the "self" variable to NULL in
spice_audio_new
This should fix https://bugzilla.redhat.com/show_bug.cgi?id=772118