Don't release images after the tail gap
This fix a hang on an image lookup that has already been remved.  It
only happens in multihead, when stressing a bit the displays.

The decoder assumed that the last added images is the tail of the
window. However, there are "gaps" in the window that will be filled by
other channels. Instead of taking the last added image as the current
up to date tail, let's take the reference to the last image before the
first gap as the up to date tail, and release from there.

The spicec code does things differently, perhaps it needs slightly
less memory at the cost of added complexity by maintaining list of
missing images and much more conditions/synchronization.