widget: factor out update_size_request() and scaling_updated()
Factor out and simplify a little the code to allow easier addition of
zoom-level property.

The "set_display" parameter for recalc_geometry() seems to be useless,
since the code was apparently trying to set the guest to the given
d->width and d->height, but reseting it to the current value, which
cancel the effect.

We should fix the problem of setting the guest resolution at
display-init time in a follow-up patch since it probably never worked
with spice-gtk.