widget: allow defining a zoom-level
Add a "zoom-level" property. Maintain the given scaling ratio when
scaling is enabled.

If scaling is disabled, this property is ignored.

For example at 50%, this allows to fit a 640x480 desktop in a 320x240
widget and when resizing it to 640x512 to have a 1280x1024 desktop.

(this feature is required by virt-viewer)