Import GNULIB's GCC warning macros
GNULIB has a helpful module 'manywarnings' which makes it easy
to turn on every single GCC warning. The general goal is that
every possible GCC warning should be enabled, except for certain
blacklisted warnings.

This imports the GNULIB m4 macros, and updates configure.ac to
use this new capability. As & when new GCC warnings are created,
the 'manywarnings.m4' can be refreshed from upstream GNULIB

* m4/manywarnings.m4, m4/warnings.m4: GNULIB warning macros
* configure.ac: Remove custom compiler warning checks
* m4/spice-compile-warnings.m4: Decide what GCC warnings to enable