Fix non-semi-seamless migration in spicy
The windows are destroyed during non-semi-seamless migrations, but the
gtk session and connected handlers remains. When a property changes
again on it, it will signal a destroyed window and lead to a crash.

The signal handler should be disconnected when the window is
destroyed. Since we have N numbers of handlers, it's easier to use
spice_signal_connect_object() helper to handle this for us by turning
spice_window structure into a basic GObject.

That GObject code could be improved, but that wasn't the goal of this
patch.