Do not grab display widget
We used to gtk_grab_add() after mouse grab to limit the mouse events
to the display. But it isn't necessary if the display has its own
window, since gdk_pointer_grab() will be limited to it.

Note the widget has the keyboard focus & is focused, so this doesn't
affect keyboard events.

However, this allows application to keep global accelerators
functionning if they want to (like virt-viewer with custom key
bindings).