Hide cursor when it is on a different screen in server mode
When the cursor shape is changed, all the cursor channels are
updated. The current code assumed that the "set" of the shape should
show the cursor, but it should stay hidden instead.

Also, when the cursor is hidden, we must invalidate its current
region to redraw display.

Fix: https://bugzilla.redhat.com/show_bug.cgi?id=804308