Handle grab-broken event
This fixes the pointer being "trapped" in the center of the screen in
server-side mouse mode. It also correctly inform the client that the
pointer/keyboard is no longer grabbed so it can adjust its UI state
accordingly (remote the "press ctrl+alt to ungrab" messages etc).

I can reproduce only with RHEVM22 host, and a RHEL6 guest, when
switching consoles.