autogen.sh: default to --enable-vala when building from git
People using autogen.sh are likely to be building from git, so may
get updates to vala files at any time. Checking for the presence of
controller.vala.stamp to decide whether vala should be enabled or not
is not very accurate since it doesn't reflect if a .vala file needs
to be regenerated or not.
It's better to always pass --enable-vala to configure, it's always
possible to disable it by using --disable-vala as an autogen.sh argument.