Fix multiple problems with URI parsing
The URI parsing was not correctly skipping over the path component
if a port number was specified using the traditional URI scheme,

eg

   spice://somehost:5900/

would result in failed parse due to the trailing '/'

The URI parsing was also not considering that the authority
component is allowed to contain '[' and ']' around the hostname.
This is used when specifying raw IPv6 addresses to remove the
parsing ambiguity wrt ':'.

eg

   spice://[::1]:5900/

Various stages of parsing also failed to report errors.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>