Clipboard clean-up and fixes for Windows client
The Windows client was getting in the way of guest copy-paste,
because when the guest was taking the clipboard grab, the agent
notifies the client, it takes the grab too, and in return
receives clipboard notification of new ownership from the client
clipboard.

Though we had a hack to check if this new client clipboard event
is caused by us, the Windows Gtk clipboard is giving 2 notifications
for some reasons.

It turned out there is a much better way than the "selfgrab" hack,
by setting ownership of the clipboard. Problem solved, and cleaner
code!

https://bugzilla.redhat.com/show_bug.cgi?id=822688