Add a desktop-integration helper class
We need to integrate closely with the desktop environment of the user in
several cases. Some examples are disabling auto-mounting when auto-usbredir
is active (rhbz#812972), and disabling the screensaver when fullscreen
(fdo#34793).

Unfortuntely these kinds of things require desktop environment specific
handling. Therefor this patch introduces a desktop-integration helper class,
which is to server as a container for all sort of desktop environment specific
functions.

For now it just supports disabling automounting under Gnome, but this will be
extended in the future.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>