Add SpiceDisplay:ready property
There are several condition to meet in order to have a widget ready to
be displayed: the monitor must exist, the area must intersect, and the
display mark must be reached. This  property will help clients to know
when the  widget display is  ready. Until now,  it was relying  on the
channel  mark signal  only,  and had  to  deal with  the  rest of  the
conditions without much help.