Enable the display early when the widget is created
The  spice_main_set_display_enabled() function  is  used  to mark  the
display/monitor  config  as  enabled.  In  order  to  simplify  client
implementation, the widget enables  the matching display automatically
when the channel mark is received. This is only for legacy reason, and
my lack  of understanding at  that time. It  could as well  be enabled
earlier, when the widget is created.  It wasn't really a good decision
to disable monitor when the mark is off, which can be toggled when the
primary surface is resize for example, and can cause some races..