usb-device-manager: do not try to connect a usb device that was removed
If a device that is asked to be shared with the guest, is unplugged out
of the machine before being redirected, then let the user know that
usbredir of that device failed (and cleanup nicely).

For Windows client, the time between request and redir is larger, as
it includes the time it takes to install the libusb driver.