build: add --enable-smartcard=auto support
Currently, when running configure with no arguments, smartcard
support is enabled by default, and configure will fail if it cannot
find libcacard. This commit adds a --enable-smartcard=auto mode to
configure which will use automatically enable smartcard support if
libcacard is available, but it will be silently disabled if libcacard
is not available. Passing --enable-smartcard will fail if libcacard
is not available. Passing --disable-smartcard will always disable
smartcard support and will not test for libcacard availability.