build: generate sym-file
When -Wl,--version-script is not available, we fallback to using
libtool --export-symbols feature, but the needed sym-file is missing
from git. This commit adds generation of the sym-file to the
update-map-file Makefile target, and adds a sym-file to git.
The linker on OSX Lion doesn't like to have non-existing symbols
specified in the symbol files it's given, so the symbols need to
be split in glib symbols and gtk symbols.