inputs: add spice_inputs_key_press_and_release()
If the server is capable of SPICE_INPUTS_CAP_SCANCODE, then we send
can send a single message with key press and release, to avoid
unwanted guest side key repeatition due to network jitter.

If the server is not capable, spice-gtk will use some compatibility
mode and send the existing DOWN and UP key events seperately.