usb: Add info message when USB dialog is empty
From rh bug #804187:
« The redirection dialog can feel a bit strange when there is no device to
redirect.

It could be useful to provide a help message indicating that there is no
device to redirect yet, and that the user can insert a USB device to
redirect, and some related guidance. »

This commit adds a "No USB devices detected" infobar in the USB
dialog below the 'Select USB devices to redirect" label.
Content could probably be improved, but this is a step in the right
direction ;)

This can be tested with
diff --git a/gtk/usb-device-widget.c b/gtk/usb-device-widget.c
index b1bf090..660ea03 100644
--- a/gtk/usb-device-widget.c
+++ b/gtk/usb-device-widget.c
@@ -220,6 +220,11 @@ static GObject *spice_usb_device_widget_constructor(
                      G_CALLBACK(device_error_cb), self);

     devices = spice_usb_device_manager_get_devices(priv->manager);
+    if (devices) {
+        g_ptr_array_unref(devices);
+        devices = NULL;
+    }
+
     if (!devices)
         goto end;